大附件问题

+++ public_models/utils.py
@@ -78,7 +78,7 @@ def get_content_type(path,path_front,file):
 
     url = '{}{}{}'.format(path, file.path, file.file_name)
     # 如果是office文件，则同路径下有pdf文件
-    if url.endswith('xls') or url.endswith('xlsx') or url.endswith('doc') or url.endswith('docx'):
+    if url.endswith('xls') or url.endswith('XLS') or url.endswith('xlsx') or url.endswith('XLSX') or url.endswith('doc') or url.endswith('DOC') or url.endswith('docx') or url.endswith('DOCX'):
         url_pdf_list = url.split('.')
         url_office_type = url_pdf_list.pop()
         url_pdf_list.append('pdf')
@@ -87,23 +87,23 @@ def get_content_type(path,path_front,file):
         url_pdf = url_pdf.replace(path, path_front)
         url = url.replace(path, path_front)
 
-        type = 'excel' if url.endswith('xls') or url.endswith('xlsx') else 'doc'
+        type = 'excel' if url.endswith('xls') or url.endswith('XLS') or url.endswith('xlsx') or url.endswith('XLSX') else 'doc'
         dict_a = content_type(type, file.file_name, url_pdf, url,file.file_caption)
 
 
 
     # 如果是图片或者是pdf
-    elif url.endswith('jpg') or url.endswith('png') or url.endswith('jpeg') or url.endswith('bmp') or url.endswith(
-            'gif') or url.endswith('pdf'):
-        type = 'pdf' if url.endswith('pdf') else 'image'
+    elif url.endswith('jpg') or url.endswith('JPG') or url.endswith('png') or url.endswith('PNG') or url.endswith('jpeg') or url.endswith('JPEG') or url.endswith('bmp') or url.endswith('BMP') or url.endswith(
+            'gif') or url.endswith('GIF') or url.endswith('pdf') or url.endswith('PDF'):
+        type = 'pdf' if url.endswith('pdf') or url.endswith('PDF') else 'image'
         url_jpg_pdf = url.replace(path, path_front)
         dict_a = content_type(type, file.file_name, url_jpg_pdf, url_jpg_pdf,file.file_caption)
 
 
     # 如果是txt或者zip
-    elif url.endswith('txt') or url.endswith('zip'):
+    elif url.endswith('txt') or url.endswith('TXT') or url.endswith('zip') or url.endswith('ZIP') or url.endswith('RAR') or url.endswith('rar'):
         url_t_z = url.replace(path, path_front)
-        type = 'txt' if url.endswith('txt') else 'zip'
+        type = 'txt' if url.endswith('txt') or url.endswith('TXT') else 'zip'
         dict_a = content_type(type, file.file_name, url_t_z, url_t_z,file.file_caption)
 
 
@@ -206,7 +206,7 @@ def move_attachment(tname_attachment,ecode):
                 file.delete()
                 url = '{}{}{}'.format(relative_path, file.path,file.file_name)
 
-                if url.endswith('doc') or url.endswith('docx') or url.endswith('xls') or url.endswith('xlsx'):
+                if url.endswith('doc') or url.endswith('DOC') or url.endswith('docx') or url.endswith('DOCX') or url.endswith('xls') or url.endswith('XLS') or url.endswith('xlsx') or url.endswith('XLSX'):
                     # 拼接相同路径下的pdf路径
                     url_list = url.split('.')
                     url_office = url_j_c_list.pop()
@@ -242,7 +242,9 @@ def move_attachment(tname_attachment,ecode):
 
                     url_x = url_x + file.file_name
 
-                    if url_x.endswith('doc') or url_x.endswith('docx') or url_x.endswith('xls') or url_x.endswith('xlsx'):
+                    if url.endswith('doc') or url.endswith('DOC') or url.endswith('docx') or url.endswith(
+                            'DOCX') or url.endswith('xls') or url.endswith('XLS') or url.endswith(
+                            'xlsx') or url.endswith('XLSX'):
                         #拼接临时路径下的pdf
                         url_j_c_list = url_j_c.split('.')
                         url_j_c_office = url_j_c_list.pop()
@@ -254,8 +256,8 @@ def move_attachment(tname_attachment,ecode):
                         url_x_office = url_x_pdf_list.pop()
                         url_x_pdf_list.append('pdf')
                         url_x_pdf = '.'.join(url_x_pdf_list)
-
-                        shutil.move(url_j_c_pdf, url_x_pdf)
+                        if os.path.exists(url_j_c_pdf):
+                            shutil.move(url_j_c_pdf, url_x_pdf)
                     shutil.move(url_j_c, url_x)
 
                     # 删除临时目录
